NONE: code cleanup, removing obsolete type casts

+++ clover-core/src/main/java/com/atlassian/clover/registry/format/CoverageSegment.java
@@ -112,7 +112,7 @@ public class CoverageSegment {
 
         int[] hitCounts = new int[(int)(covByteLen / 4)];
 
-        int hitCountBufferSize = (int) Math.min((long) Integer.MAX_VALUE, covByteLen);
+        int hitCountBufferSize = (int) Math.min(Integer.MAX_VALUE, covByteLen);
         ByteBuffer hitCountsBuffer = ByteBuffer.allocate(hitCountBufferSize);
         //Read each page of bytes
         for (long curByteCount = 0; curByteCount < covByteLen; curByteCount += Integer.MAX_VALUE) {
@@ -130,7 +130,7 @@ public class CoverageSegment {
         final long startPos = channel.position();
 
         int[] hitCountsVal = hitCounts.get();
-        int hitCountByteBufferSize = (int)Math.min((long) Integer.MAX_VALUE, ((long)hitCountsVal.length) * 4);
+        int hitCountByteBufferSize = (int)Math.min(Integer.MAX_VALUE, ((long)hitCountsVal.length) * 4);
         final ByteBuffer hitCountByteBuffer = ByteBuffer.allocate(hitCountByteBufferSize);
         final IntBuffer hitCountIntBuffer = hitCountByteBuffer.asIntBuffer();
         //Write each page of bytes
